Fixed comments for Box class
