--- src/homework_4/Ani_Barseghyan/chapter_6/BoxDemo2/Box.java
@@ -1,7 +1,0 @@
-package homework_4.Ani_Barseghyan.chapter_6.BoxDemo2;
-
-public class Box {
-    double width;
-    double height;
-    double depth;
-}
