--- src/homework_4/Ani_Barseghyan/chapter_6/BoxDemo3/Box.java
@@ -1,12 +1,0 @@
-package homework_4.Ani_Barseghyan.chapter_6.BoxDemo3;
-
-public class Box {
-    double width;
-    double height;
-    double depth;
-
-    void volume() {
-        System.out.print("Volume is ");
-        System.out.println(width * height * depth);
-    }
-}
