--- src/homework_4/Ani_Barseghyan/chapter_6/BoxDemo4/Box.java
@@ -1,11 +1,0 @@
-package homework_4.Ani_Barseghyan.chapter_6.BoxDemo4;
-
-public class Box {
-    double width;
-    double height;
-    double depth;
-
-    double volume() {
-        return width * height * depth;
-    }
-}
