+++ src/homework_4/Ani_Barseghyan/chapter_6/BoxDemo6.java
@@ -1,18 +1,15 @@
-package homework_4.Ani_Barseghyan.chapter_6.BoxDemo5;
+package homework_4.Ani_Barseghyan.chapter_6;
 
-public class BoxDemo5 {
+public class BoxDemo6 {
     public static void main(String[] args) {
         Box myBox1 = new Box();
         Box myBox2 = new Box();
         double vol;
 
-        myBox1.setDim(10, 20, 15);
-        myBox2.setDim(3, 6, 9);
-
-        vol = myBox1.volume();
+        vol = myBox1.volume2();
         System.out.println("Volume is " + vol);
 
-        vol = myBox2.volume();
+        vol = myBox2.volume2();
         System.out.println("Volume is " + vol);
     }
 }
