--- src/homework_4/Ani_Barseghyan/chapter_6/BoxDemo7/Box.java
@@ -1,17 +1,0 @@
-package homework_4.Ani_Barseghyan.chapter_6.BoxDemo7;
-
-public class Box {
-    double width;
-    double height;
-    double depth;
-
-    Box(double w, double h, double d) {
-        width = w;
-        height = h;
-        depth = d;
-    }
-
-    double volume() {
-        return width * height * depth;
-    }
-}
